requirements.txt
python-dotenv
transformers
torch
sentence-transformers

